refactor: separate getCategoryBgcolor as helper
