+++ src/containers/recordCard/content/Moment.tsx
@@ -1,6 +1,7 @@
 import React, { useContext } from "react";
 import Image from "next/image";
 import { DataContext } from "../SingleCardLayout";
+import getCategoryBgcolor from "@/common/helpers/getCategoryBgcolor";
 
 interface MomentDataType {
   TITLE: string;
@@ -13,23 +14,6 @@ const Moment: React.FC = () => {
     return null;
   }
   const { category, content, photo, desc } = data;
-
-  const getCategoryBgcolor = (prop: string) => {
-    switch (prop) {
-      case "行為表現":
-        return "bg-[#F9E6FF]";
-      case "驚喜":
-        return "bg-[#FFF5CF]";
-      case "生活習慣":
-        return "bg-[#FFE9EC]";
-      case "社交":
-        return "bg-[#D5F0FF]";
-      case "技能":
-        return "bg-[#E0FFDF]";
-      default:
-        return "";
-    }
-  };
 
   const momentData: MomentDataType[] = [
     {
